Smart-Bug-Triage-Agent/requirements.txt
# Core Dependencies
sqlalchemy>=1.4.0
requests>=2.28.0
pika>=1.3.0
python-dotenv>=0.19.0

# NLP and Machine Learning
spacy>=3.4.0
scikit-learn>=1.1.0
pandas>=1.4.0
numpy>=1.21.0

# Web Framework (for APIs and webhooks)
fastapi>=0.85.0
uvicorn>=0.18.0

# Database Drivers
psycopg2-binary>=2.9.0  # PostgreSQL
# sqlite3 is built into Python - no need to install

# Monitoring and Logging
prometheus-client>=0.14.0
structlog>=22.1.0

# GitHub Integration
PyGithub>=1.55.0

# Jira Integration (optional)
jira>=3.4.0

# Slack Integration (optional)
slack-sdk>=3.18.0

# Email Integration (optional)
sendgrid>=6.9.0

# Testing (optional)
pytest>=7.1.0
pytest-asyncio>=0.19.0
pytest-cov>=3.0.0

# Development Tools (optional)
black>=22.6.0
flake8>=5.0.0
mypy>=0.971